// src/app/core/models/ViewModel/vm-maintenance-contract-dues.ts
export class VMMaintenanceContractDues {
    id: any;
    contractNumber: any;
    maintenanceContractId: any;
    dueAmount: any;
    typeId: any;
    dueStartDate: any;
    dueEndDate: any;
    isEntryGenerated!: boolean;
    isInvoiced!: boolean;
    contractSettingId:any;
    paid: any;
    typeNameAr!: string;
    typeNameEn!: string;
    supplierId: any;
    supplierNameAr!: string;
    supplierNameEn!: string;
    contractId: any;
    notes!:string;
    contractStatus!:number;
    contractSettingNameAr!:string;
    contractSettingNameEn!:string;

}